+++ features/Context/AssertionContext.php
@@ -203,6 +203,7 @@ class AssertionContext extends RawMinkContext
     /**
      * @param TableNode $tableNode
      *
+     * @Then /^I should see a dialog with the following content:$/
      * @Then /^I should see a confirm dialog with the following content:$/
      *
      * @throws ExpectationException