+++ spec/Pim/Bundle/CatalogBundle/Doctrine/Common/Remover/CategoryRemoverSpec.php
@@ -3,6 +3,7 @@
 namespace spec\Pim\Bundle\CatalogBundle\Doctrine\Common\Remover;
 
 use Akeneo\Component\StorageUtils\Remover\RemovingOptionsResolverInterface;
+use Akeneo\Component\StorageUtils\Saver\BulkSaverInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Event\CategoryEvents;
@@ -16,9 +17,10 @@ class CategoryRemoverSpec extends ObjectBehavior
     function let(
         ObjectManager $objectManager,
         RemovingOptionsResolverInterface $optionsResolver,
-        EventDispatcherInterface $eventDispatcher
+        EventDispatcherInterface $eventDispatcher,
+        BulkSaverInterface $productSaver
     ) {
-        $this->beConstructedWith($objectManager, $optionsResolver, $eventDispatcher);
+        $this->beConstructedWith($objectManager, $optionsResolver, $eventDispatcher, $productSaver);
     }
 
     function it_is_a_remover()
@@ -29,10 +31,13 @@ class CategoryRemoverSpec extends ObjectBehavior
     function it_dispatches_an_event_when_removing_a_category(
         $eventDispatcher,
         $objectManager,
+        $optionsResolver,
+        $productSaver,
         CategoryInterface $category,
         ProductInterface $product1,
         ProductInterface $product2
     ) {
+        $optionsResolver->resolveRemoveOptions(['flush' => true])->willReturn(['flush' => true]);
         $category->isRoot()->willReturn(false);
         $category->getProducts()->willReturn([$product1, $product2]);
         $product1->removeCategory($category)->shouldBeCalled();
@@ -44,8 +49,18 @@ class CategoryRemoverSpec extends ObjectBehavior
         )->shouldBeCalled();
 
         $objectManager->remove($category)->shouldBeCalled();
+        $objectManager->flush()->shouldBeCalled();
 
-        $this->remove($category, ['flush' => false]);
+        $productSaver->saveAll(
+            [$product1, $product2],
+            [
+                'flush'       => true,
+                'recalculate' => false,
+                'schedule'    => false,
+            ]
+        )->shouldBeCalled();
+
+        $this->remove($category, ['flush' => true]);
     }
 
     function it_dispatches_an_event_when_removing_a_tree(
@@ -63,7 +78,7 @@ class CategoryRemoverSpec extends ObjectBehavior
 
         $objectManager->remove($tree)->shouldBeCalled();
 
-        $this->remove($tree, ['flush' => false]);
+        $this->remove($tree, ['flush' => true]);
     }
 
     function it_throws_exception_when_remove_anything_else_than_a_category()